gemini-balancer/pyproject.toml
[project]
name = "gemini-balancer"
version = "0.1.0"
description = "Simple Gemini API model balancer with user-configured rate limits."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
]

[project.scripts]
gemini-balancer = "gemini_balancer.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
